Modified poc.py
poc.py allows the program check the emotions of the person inside a video and saves the results as a .jpg and .json file.
changes: .jpg and .json files are now saved to the directory once every 100 frames with the information for a specific frame. .jpg files show the entire frame while the .json files contain a dictionary with the frame number and the length of the rectangles for each emotion.